# frontend/public/demo-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverrideSuggestion": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "correction_count": {
            "title": "Correction Count",
            "type": "integer"
          },
          "last_corrected": {
            "title": "Last Corrected",
            "type": "string"
          },
          "suggested_category": {
            "title": "Suggested Category",
            "type": "string"
          }
        },
        "required": [
          "company",
          "suggested_category",
          "correction_count",
          "last_corrected"
        ],
        "title": "OverrideSuggestion",
        "type": "object"
      },
      "OverrideSuggestionsResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/OverrideSuggestion"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "suggestions",
          "count"
        ],
        "title": "OverrideSuggestionsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal finance dashboard API. Ingests bank transaction emails, categorizes them (with optional OpenAI enrichment), and exposes transactions, budgets, summaries, and AI spending insights over a consistent JSON HTTP interface.\n\n**Storage:** dual-backend \u2014 DynamoDB (AWS) or SQLite (local/demo). Selection is driven by `data/config.json`. All endpoints return identical shapes regardless of backend.\n\n**Versioning:** routes are prefixed with `/api/v1/`. Breaking changes will ship as `/api/v2/` rather than in-place.\n\n**Demo:** this is a read-only snapshot of the Tidings demo journal \u2014 the fictional Mira Lin Chen household, covering 2025-05 through 2026-03. All data is invented. Any request other than GET, HEAD, or OPTIONS returns 405; self-host Tidings for a writable journal with bearer-token auth.",
    "title": "Tidings demo API (read-only)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/overrides/suggestions": {
      "get": {
        "description": "Analyze manual category corrections to suggest new override rules.\n\nAlgorithm: find companies with 1+ manual corrections to the same target\ncategory that don't already have an override.",
        "operationId": "getOverrideSuggestions",
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 24,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideSuggestionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest new override rules from manual category corrections",
        "tags": [
          "overrides"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gettidings.com/demo"
    }
  ]
}